Reworked how colours are drawn.
The old algorithm drew colours depending on the "depth" (the number of iterations before the program realised a point was in the Mandelbrot set), but it took depth as a proportion of the maximum depth (exceeding the maximum depth assumes that point indeed *is* in the Mandelbrot set, thus the pixel is left as was). The new algorithm is absolute, so changing non-colour related parameters does not affect the colour scheme, and colours can be adjusted in isolation much more easily.

There is also now a "maximum" colour, if the depth goes beyond the colour gradients defined in the length to colour function, then it goes to a maximum colour. This is best left as the colour just on the edge of the previous band for better visual smoothness, but this is not required/forced.