Added rolling colours
The colour sequence used to be that the more iterations it took to be clear a point was *not* in the Mandelbrot set went as such: black -> blue -> green -> yellow -> red -> white, anything above a certain number of iterations was just white. Now it goes black -> blue -> green -> yellow -> red -> magenta -> blue -> green etc. It just keeps going with higher complexities and more detail if the hardware allows.